Move the selection computation code into a separate file
